x86_64 done, multiboot2 done, code refactoring blah blah no one cares

+++ src/drivers/keyboard.c
@@ -130,5 +130,5 @@ void keyboard_init(void) {
     uint8_t status = inb(KEYBOARD_CMD_PORT);
     terminal_printf(PRINT_DEBUG, "Keyboard controller status: 0x%x\n", status);
 
-    idt_set_entry(33, (uint32_t)keyboard_handler_stub, 0x08, 0x8E);
+    idt_set_entry(33, (uint64_t)keyboard_handler_stub, 0x08, 0x8E);
 }
